Vamos!! He conseguido crear la tabla clientes que descienda de CUENTAS pero sin que se use la tabla cuentas
