+++ src/main/java/org/springframework/samples/petclinic/model/Cuenta.java
@@ -4,6 +4,7 @@ import java.time.LocalDate;
 
 import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.MappedSuperclass;
 import javax.persistence.Table;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Size;
@@ -15,8 +16,7 @@ import lombok.EqualsAndHashCode;
 
 @Data
 @EqualsAndHashCode(callSuper=false)
-@Entity
-@Table(name = "Cuentas")
+@MappedSuperclass
 public class Cuenta extends BaseEntity {
 	
 	@Column(name = "nombre")
